+++ Assets/Tests/EditMode/SimulationRunnerFlowTests.cs
@@ -153,7 +153,7 @@ namespace Tests.EditMode
         }
 
         private void FillAllVacuum() { for (int y = 0; y < _grid.Height; y++) for (int x = 0; x < _grid.Width; x++) SetCell(x, y, VacuumId, 0); }
-        private void SetCell(int x, int y, byte elementId, int mass, short temperature = 0) { ref SimCell c = ref _grid.GetCellRef(_grid.ToIndex(x, y)); c = new SimCell(elementId, mass, temperature, SimCellFlags.None); }
+        private void SetCell(int x, int y, byte elementId, int mass, float temperature = 0f) { ref SimCell c = ref _grid.GetCellRef(_grid.ToIndex(x, y)); c = new SimCell(elementId, mass, temperature, SimCellFlags.None); }
         private int SumMassOfElement(byte elementId) { int t = 0; for (int i = 0; i < _grid.Length; i++) if (_grid.GetCellByIndex(i).ElementId == elementId) t += _grid.GetCellByIndex(i).Mass; return t; }
         private int SumEntityMass(byte elementId) { int t = 0; var e = _runner.FallingEntities.ActiveEntities; for (int i = 0; i < e.Count; i++) if (e[i].ElementId == elementId) t += e[i].Mass; return t; }
